Renamed and Added functionality to test for edge cases in the api interface

+++ server-excercises/todoApp/public/apiInterface.js
@@ -10,6 +10,8 @@ app.api = {
   },
 
   updateAll: function (status) {
+    const isValid = this.checkValidity(1, 'test', status)
+    if (isValid !== true) return isValid
     const data = {
       status: status
     }
@@ -21,12 +23,12 @@ app.api = {
         'Content-Type': 'application/json'
       }
     })
-      .catch(function (err) {
-        console.log(err)
-      })
   },
 
   updateTask: function (id, description, status) {
+    const isValid = this.checkValidity(id, description, status)
+    if (isValid !== true) return isValid
+
     const data = {
       status: status,
       task: description
@@ -39,21 +41,36 @@ app.api = {
         'Content-Type': 'application/json'
       }
     })
-      .catch(function (err) {
-        console.log(err)
-      })
   },
 
   insertTask: function (description) {
+    const isValid = this.checkValidity(1, description, true)
+    if (isValid !== true) return isValid
+
+    this.checkValidity(1, description, false)
     return fetch(`/write/${description}`, { method: 'post' })
   },
 
   deleteTask: function (id) {
+    const isValid = this.checkValidity(id, 'description', 'status')
+    if (isValid !== true) return isValid
+
+    this.checkValidity(id, 'string', false)
     return fetch(`/destroy/${id}`, { method: 'delete' })
   },
 
   deleteCompleted: function () {
+    this.checkValidity(1, 'string', status)
     return fetch(`/destroycompleted`, { method: 'delete' })
+  },
+  checkValidity: function (id, description, status) {
+    if (typeof id !== 'number') return `id-${id} is not a number`
+
+    if (typeof description !== 'string') return `description-${description} is not a string`
+
+    if (typeof status !== 'boolean') return `status given is not a boolean`
+
+    return true
   }
 }
